Make sure that operator % throws an exception when the ride-hand-side argument is 0
fixes #8089

+++ lib/config/expression.cpp
@@ -159,12 +159,12 @@ Value DivideExpression::DoEvaluate(ScriptFrame& frame, DebugHint *dhint) const
 
 Value ModuloExpression::DoEvaluate(ScriptFrame& frame, DebugHint *dhint) const
 {
-	return (long)m_Operand1->Evaluate(frame) % (long)m_Operand2->Evaluate(frame);
+	return m_Operand1->Evaluate(frame) % m_Operand2->Evaluate(frame);
 }
 
 Value XorExpression::DoEvaluate(ScriptFrame& frame, DebugHint *dhint) const
 {
-	return (long)m_Operand1->Evaluate(frame) ^ (long)m_Operand2->Evaluate(frame);
+	return m_Operand1->Evaluate(frame) ^ m_Operand2->Evaluate(frame);
 }
 
 Value BinaryAndExpression::DoEvaluate(ScriptFrame& frame, DebugHint *dhint) const
